+++ scrapy_poet/middleware.py
@@ -13,8 +13,12 @@ from twisted.internet.defer import inlineCallbacks
 from scrapy.utils.misc import create_instance, load_object
 
 from .api import DummyResponse
+from .page_input_providers import (
+    HttpClientProvider,
+    HttpResponseProvider,
+    PageParamsProvider,
+)
 from .overrides import OverridesRegistry
-from .page_input_providers import HttpResponseProvider
 from .injection import Injector
 
 
@@ -22,7 +26,9 @@ logger = logging.getLogger(__name__)
 
 
 DEFAULT_PROVIDERS = {
-    HttpResponseProvider: 500
+    HttpResponseProvider: 500,
+    HttpClientProvider: 600,
+    PageParamsProvider: 700,
 }
 
 InjectionMiddlewareTV = TypeVar("InjectionMiddlewareTV", bound="InjectionMiddleware")